pyproject.toml
[tool.poetry]
name = "rosnik"
version = "0.0.8"
description = "ROSNIK Python SDK"
authors = ["Nick DiRienzo <nick@rosnik.ai>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
# Target 3.8+ since that's receiving patches through 2025
# Don't support 4 yet.
python = ">=3.10,<4"
openai = {version = "^0.27.6", optional = true}
ulid-py = "^1.1.0"

[tool.poetry.group.static.dependencies]
bandit = "^1.7"
safety = "^2.3"
black = "^23.3.0"
mypy = "^1.5.1"
ruff = "^0.0.267"
poetry-dotenv-plugin = "^0.1.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
pytest-recording = "^0.13.0"

[tool.poetry.extras]
openai = ["openai"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100